+++ src/server/app-rest-media.ts
@@ -43,26 +43,30 @@ export default class RESTMedia {
                 const id2name = (id: string) =>
                     id.replaceAll(/(^|-)(\S)/g, (_, prefix, char) =>
                         (prefix ? " " : "") + char.toUpperCase())
-                const files = await glob("**/*.{png,jpg,gif,mp4,webm}", { cwd: mediaDir })
+                const files = await glob("**/*.{png,jpg,gif,smp4,swebm,mp4,webm}", { cwd: mediaDir })
                 for (const file of files) {
                     if (media.includes(file))
                         continue
                     let m: ReturnType<typeof file.match>
-                    if ((m = file.match(/^(.+)\.(png|jpg|gif|mp4|webm)$/)) !== null) {
+                    if ((m = file.match(/^(.+)\.(png|jpg|gif|smp4|mp4|swebm|webm)$/)) !== null) {
                         let type = ""
-                        if      (m[2] === "png")  type = "Image/PNG"
-                        else if (m[2] === "jpg")  type = "Image/JPEG"
-                        else if (m[2] === "gif")  type = "Image/GIF"
-                        else if (m[2] === "mp4")  type = "Video/MP4"
-                        else if (m[2] === "webm") type = "Video/WebM"
-                        const loop = (m[1].match(/-loop$/) !== null)
+                        if      (m[2] === "png")   type = "Image/PNG"
+                        else if (m[2] === "jpg")   type = "Image/JPEG"
+                        else if (m[2] === "gif")   type = "Image/GIF"
+                        else if (m[2] === "mp4")   type = "Video/MP4"
+                        else if (m[2] === "smp4")  type = "Video/sMP4"
+                        else if (m[2] === "webm")  type = "Video/WebM"
+                        else if (m[2] === "swebm") type = "Video/sWebM"
+                        const loop  = (m[1].match(/-loop$/) !== null)
+                        const stack = (m[2] === "smp4" || m[2] === "swebm")
                         result.media.push({
                             id:        m[1],
                             name:      id2name(m[1].replace(/^.+?\//, "")),
                             group:     m[1].replace(/^[^/]+$/, "").replace(/^(.+)\/.+$/, "$1"),
                             texture:   `${mediaURL}/${m[1]}.${m[2]}`,
                             type,
-                            loop
+                            loop,
+                            stack
                         })
                     }
                 }